+++ assets/scripts/routes/settings.js
@@ -12,7 +12,7 @@ import {
 import { handleErrors } from '../utils'
 import Settings from '../components/screens/Settings.svelte'
 import page from 'page'
-import { deleteFileAndSaveChanges } from '../actions/edition'
+import { deleteFileAndSaveChanges } from '../actions/file'
 
 const blogMdContent = `---
 layout: default
